move docker image building to be a global autouse fixture to make sure it runs before tests that use output
